Updating WebInkEnhancements explainer to include canvas and improvement metric (#195)
* Adding clip rect size and improvement metrics

Updating the IDL for the presenter to now accept the canvas (or whatever is being drawn to). This will allow us to get the size of the canvas to set the clip rect of the presentation area.

Also updating to return a metric to the web developer to make the decision on if they want to use this API based on the expected performance gains.