Introduce SOTD
Expected format for future explainers' Status of this Document sections to help the community always find the latest version of the document. Prior to transition to another venue, document status should be **Active**, and "Current venue" should read "Expected venue" so that reviewers know where we expect this document to go next (if known). "Current version" can have a value of "this document".