[Title Bar Customization] Swap "caption controls" for "window controls" (#248)
The term "caption controls" describes the caption buttons: minimize, maximize, close. This terminology may be familiar for Windows developers, but is confusing for those unfamiliar with it, leading people to think this is somehow referring to closed captions. This change swaps the term "caption controls" with "window controls" for clarity.