Add open-ui links to 'Standardized control UI anatomy' section (#363)
Add open-ui links to the 'Standardized control UI anatomy, parts, and behavior' section. Open-UI is already referenced in the 'Incremental Approach' section but it seems best to include it in the section about standardized anatomies as well since Open-UI is the venue for driving that process.

Context: #361 made me realize that Open UI wasn't really mentioned in the appropriate place.