[Crash reports] Add Security & Privacy Self-Review Questionnaire (#823)
This PR includes the completed Security & Privacy self-review questionnaire as per the requirement of the Chrome Entry. The questionnaire addresses various aspects of the feature, including the exposure of information, handling of sensitive data, persistence of state, interaction with the underlying platform, and more.

The questionnaire is designed to ensure that all security and privacy considerations are thoroughly reviewed and addressed. It follows the format provided by the W3C: https://www.w3.org/TR/security-privacy-questionnaire/

The responses in the questionnaire provide a comprehensive overview of how the feature interacts with user data and the browser environment.