Update Open Issues Declarative CSS Modules Explainer (#1132)
One of the open issues the [TAG review](https://github.com/w3ctag/design-reviews/issues/1000#issuecomment-3162121691) brought up is regarding import maps:

> Was it ever intentional that <script type="importmap"> works inside a shadow root? It'd be good to check with that feature's designers (e.g. @domenic) to ensure that encouraging its use there won't break any assumptions or encourage any footguns.

This slipped though the cracks on the last spec update from TAG review feedback, so this PR adds it, nearly verbatim, to the Open Issues section. 

This also updates language of the other open issue regarding a pending fetch with the same module specifier to prefer to be safer and immediately fail, as we discussed earlier today.