[CSS Modules] Adding "one and done" suggestions to Open Issues (#1146)
Per our conversation this morning, this change adds the "one and done" open issue (along with a suggestion to address it) in the Open Issues section. It also adds an open issue for how the `LinkStyle` interface should be handled and adds Steve Orvell to the References and acknowledgements section. Finally, it removes the open issue regarding Import Maps in shadow roots, as Domenic gave a response to that, so it's no longer an open issue.